+++ index.html
@@ -230,6 +230,6 @@
     </div>
   </template>
 
-  <script type="module" src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
